back/requirements.txt
asgiref==3.5.2
astroid==2.12.9
autopep8==1.7.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.5.1
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
idna==3.3
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.9.1
pylint==2.15.2
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
wrapt==1.14.1
